# Avoid unnecessary `Dictionary` conversions in `GDScriptInstance::validate_property`