Updated `GDScriptInstance::validate_property` to only convert `PropertyInfo` to `Dictionary` if `_validate_property` function is found.

This can make `Node.duplicate` around 4-5x faster for nodes that have gdscript attached, when the script doesn't override `_validate_property`.  Compared with scripts below:

TestClass:
```gdscript
class_name TestClass extends Node2D

var a
var b
var c
var d

#func _validate_property(property: Dictionary) -> void:
#	pass
```

Main Scene:
```gdscript
extends Node

func _ready():
	var node := TestClass.new()
	var start := Time.get_ticks_msec()
	for i in 1000:
		add_child(node.duplicate())
	var end := Time.get_ticks_msec()
	print("%dms" % (end - start))
```

Old:
```
139ms
238ms (_validate_property uncommented)
```

New:
```
28ms
239ms (_validate_property uncommented)
```

<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://docs.godotengine.org/en/latest/contributing/development/code_style_guidelines.html
-->