@dalexeev 2024-10-18
* This probably fixes/improves #97829.